170947605-ft(thunk):add re-usable thunk - make thunk reusable [Finishes #170947605]
